Update to the latest Pathfinder
Pathfinder is now available as jar files pre-compiled.
This includes the JNI pre-compiled for different platforms.

There's a POM but I'm not sure how to integrate with Gradle.

This adds the dylibs (for osx) and the jar files. These should be loaded from POM but I don't know how to do that.